Cambiado el mapeo. Ahora usamos automapping y generacion automatica del esquema de la DB. Renombradas las entidades, ahora terminan todas en Entity. Falta agregar mas entidades y agregar las caracteristicas adicionales de mapeo que queramos.

+++ glimpse/MailInterfaces/AccountInterface.cs
@@ -28,7 +28,7 @@ namespace Glimpse.MailInterfaces
         }
 
 
-        public MailCollection getMailsFromHigherThan(string mailbox, int lastUID)
+        public MailCollection getMailsFromHigherThan(string mailbox, Int64 lastUID)
         {
             return this.myFetcher.GetMailDataFromHigherThan(mailbox, lastUID);
         }
